+++ src/shared_types.rs
@@ -1,7 +1,7 @@
 #[derive(Clone, Copy, Debug)]
 pub(crate) struct ChunkRange {
-    pub(crate) start: u32,
-    pub(crate) end: u32,
+    pub(crate) start: u64,
+    pub(crate) end: u64,
 }
 
 pub(crate) type ByteCount = u64;
